Separated CAS capabilities from UI facing Function class and moved them into the InterpretedFunction class; Renamed InterpretedFunction class to Expression since this name is more concise & accurate.
