+++ src/jmc/cas/Assets.java
@@ -1,0 +1,12 @@
+package jmc.cas;
+
+/**
+ * Created by Jiachen on 3/2/18.
+ * Assets
+ */
+public interface Assets {
+    String DIGITS = "0123456789.";
+    String SYMBOLS = "()+-*/^<>";
+    String VARS = "abcdfghjklmnopqrstuvwxyz";
+    String LETTERS = "abcdefghijklmnopqrstuvwxyz";
+}
