--- src/tests/InconsistencyDetection.java
@@ -1,13 +1,0 @@
-package tests;
-
-import jmc.Function;
-
-/**
- * Created by Jiachen on 21/05/2017.
- */
-public class InconsistencyDetection {
-    public static void main(String[] args){
-        Function function = Function.interpret("sec<x>^tanh<x>");
-
-    }
-}
